Virtual terminal progress.
vterm now contains its own thread safe version of the screen buffer. The old screen buffer is to be deleted.

+++ vterm/screen_buffer.h
@@ -40,25 +40,25 @@ namespace vterm {
 
 		/** Returns the number of the rows of the screenbuffer. 
 		 */
-		size_t rows() const {
+		unsigned rows() const {
 			return rows_;
 		}
 
 		/** Returns the number of columns of the screenbuffer. 
 		 */
-		size_t cols() const {
+		unsigned cols() const {
 			return cols_;
 		}
 
 		/** Returns the appropriate cell.
 		 */
-		Cell const & at(size_t col, size_t row) const {
+		Cell const & at(unsigned col, unsigned row) const {
 			ASSERT(col < cols_ && row < rows_) << "Indices " << col << ";" << row << " out of bounds " << cols_ << ";" << rows_;
 			return cells_[row * cols_ + col];
 		}
 		/** Returns the appropriate cell.
 		 */
-		Cell & at(size_t col, size_t row) {
+		Cell & at(unsigned col, unsigned row) {
 			ASSERT(col < cols_ && row < rows_) << "Indices " << col << ";" << row << " out of bounds " << cols_ << ";" << rows_;
 			return cells_[row * cols_ + col];
 		}
@@ -67,7 +67,7 @@ namespace vterm {
 
 		    TODO check concurrency and synchronization of this so that resize does not clash with the writers to the buffer
 		 */
-		void resize(size_t cols, size_t rows) {
+		void resize(unsigned cols, unsigned rows) {
 			if (cols == cols_ && rows_ == rows_)
 				return;
 			// delete old cells
@@ -79,8 +79,8 @@ namespace vterm {
 		}
 
 	private:
-		size_t rows_;
-		size_t cols_;
+		unsigned rows_;
+		unsigned cols_;
 
 		Cell * cells_ = nullptr;
